clean up text tokens long

+++ palme/model.py
@@ -117,8 +117,7 @@ class PALME(nn.Module):
     def forward(self, text_tokens, images):
         try:
             # if text_tokens.dtype != torch.long:
-            #     text_tokens = text_tokens.long()
-                
+            text_tokens = text_tokens.long()
             images = self.vit_model(pixel_values=images)["last_hidden_state"]
             print(images.shape)
             images = self.perceive(images).squeeze(1)
